fix: use global component - input, list
